-apply patch from Matthias, didn't notice any problems, has been tested on Windows since a long time -increase version number to 0.9.88 -supply the _AUTOMOC4_KDE4_PRE_TARGET_HANDLING() and _AUTOMOC4_KDE4_POST_TARGET_HANDLING() macros, so they are not anymore overriden by the ones in kdelibs/cmake/modules/KDE4Macros.cmake. This brings the actual change in cmake handling to KDE.
This reduces the time for "make plasma" in kdelibs from 5.1 s to 4.5 s on my machine if there is nothing to build.
It also makes automoc4 easier to understand (and maintain).
It increases cmake time somewhat, since for each target not also a <target>_automoc is created.

Alex

CCMAIL: kretz@kde.org
CCMAIL: buildsystem@kde.org



svn path=/trunk/kdesupport/automoc/; revision=905502