Fix missing include dirs (current source and build dir) when moc is run.
Apparently these dirs are not in the directory property "INCLUDE_DIRECTORIES",
so they need to be added by hand.

This fixes compilation of kdelibs/kdewebkit with KDE_NO_DEPRECATED=TRUE:
moc didn't find kdewebkit_export.h (and silently ignored it), so kdemacros.h
was not included, and KDE_NO_DEPRECATED was not set, leading to signals
only in the moc file and not in the header file, at compile time.

Alex: is there a way to only do this when CMAKE_INCLUDE_CURRENT_DIR is on?
CCMAIL: neundorf@kde.org